update contract sdk, update contract addresses and respective handlers
