+++ package.json
@@ -43,8 +43,8 @@
     "typescript": "^4.1.3"
   },
   "dependencies": {
-    "@subql/contract-sdk": "0.10.4-2",
-    "@subql/frontier-evm-processor": "^0.1.2",
+    "@subql/contract-sdk": "^0.11.1-0",
+    "@subql/frontier-evm-processor": "^0.1.3-0",
     "@types/pino": "^7.0.5",
     "@types/validator": "^13.7.1",
     "bs58": "^4.0.1",
