+++ src/mappings/staking.ts
@@ -11,7 +11,6 @@ import {
   UnbondRequestedEvent,
   UnbondWithdrawnEvent,
   UnbondCancelledEvent,
-  SetCommissionRateEvent,
 } from '@subql/contract-sdk/typechain/Staking';
 import assert from 'assert';
 import { Delegation, Withdrawl, Indexer, WithdrawalStatus } from '../types';
@@ -29,7 +28,6 @@ import {
   updateMaxUnstakeAmount,
 } from './utils';
 import { FrontierEvmEvent } from '@subql/frontier-evm-processor';
-import { createIndexer } from './utils';
 import { CreateWithdrawlParams } from '../interfaces';
 
 const { ONGOING, CLAIMED, CANCELLED } = WithdrawalStatus;
@@ -236,41 +234,4 @@ export async function handleWithdrawCancelled(
 
     await reportException('handleWithdrawCancelled', exception, event);
   }
-}
-
-export async function handleSetCommissionRate(
-  event: FrontierEvmEvent<SetCommissionRateEvent['args']>
-): Promise<void> {
-  logger.info('handleSetCommissionRate');
-  assert(event.args, 'No event args');
-
-  const address = event.args.indexer;
-  const eraManager = EraManager__factory.connect(
-    ERA_MANAGER_ADDRESS,
-    new FrontierEthProvider()
-  );
-
-  const lastEvent = `handleSetCommissionRate:${event.blockNumber}`;
-  let indexer = await Indexer.get(address);
-
-  if (!indexer) {
-    indexer = await createIndexer({
-      address,
-      active: true,
-      lastEvent,
-      createdBlock: event.blockNumber,
-    });
-  }
-
-  indexer.commission = await upsertEraValue(
-    eraManager,
-    indexer.commission,
-    event.args.amount.toBigInt(),
-    'replace',
-    // Apply instantly when era is -1, this is an indication that indexer has just registered
-    indexer.commission.era === -1
-  );
-  indexer.lastEvent = `handleSetCommissionRate:${event.blockNumber}`;
-
-  await indexer.save();
 }
